+++ src/test/java/racingcar/util/WinnerDecisionTest.java
@@ -1,10 +1,11 @@
-package racingcar.model;
+package racingcar.util;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.List;
 import org.junit.jupiter.api.Test;
+import racingcar.model.Car;
 
 class WinnerDecisionTest {
     @Test
